# limb: always inline bitand