--- a/src/limb/bit_and.rs
+++ b/src/limb/bit_and.rs
@@ -5,6 +5,7 @@ use core::ops::BitAnd;
 
 impl Limb {
     /// Calculates `a & b`.
+    #[inline(always)]
     pub const fn bitand(self, rhs: Self) -> Self {
         Limb(self.0 & rhs.0)
     }
@@ -13,6 +14,7 @@ impl Limb {
 impl BitAnd for Limb {
     type Output = Limb;
 
+    #[inline(always)]
     fn bitand(self, rhs: Self) -> Self::Output {
         self.bitand(rhs)
     }