@haslersn 2022-08-08
Though I'm not sure why `sub_mod` is still so much slow than `sub_mod_special` from #108. Here's the assembly of `sub_mod_special` for comparison, using the same `const` modulus (just represented differently for `sub_mod_special`):

<details>
<summary>Assembly generated for U256::sub_mod_special from #108</summary>

```asm
crypto_bigint::uint::sub_mod::<impl crypto_bigint::uint::UInt<_>>::sub_mod_special:
 xor     r11d, r11d
 mov     r8, qword, ptr, [rsi]
 sub     r8, qword, ptr, [rdx]
 mov     r10d, 0
 sbb     r10, r10
 mov     r9, qword, ptr, [rsi, +, 8]
 sar     r10, 63
 sub     r9, qword, ptr, [rdx, +, 8]
 mov     ecx, 0
 sbb     rcx, rcx
 add     r9, r10
 adc     rcx, r10
 mov     r10, qword, ptr, [rsi, +, 16]
 sar     rcx, 63
 sub     r10, qword, ptr, [rdx, +, 16]
 mov     eax, 0
 sbb     rax, rax
 add     r10, rcx
 adc     rax, rcx
 mov     rsi, qword, ptr, [rsi, +, 24]
 sar     rax, 63
 sub     rsi, qword, ptr, [rdx, +, 24]
 mov     ecx, 0
 sbb     rcx, rcx
 add     rsi, rax
 adc     rcx, rax
 and     ecx, 6292479
 sub     r8, rcx
 sbb     r11, r11
 sar     r11, 63
 add     r9, r11
 adc     r11, 0
 sar     r11, 63
 add     r10, r11
 adc     r11, 0
 sar     r11, 63
 add     r11, rsi
 mov     qword, ptr, [rdi], r8
 mov     qword, ptr, [rdi, +, 8], r9
 mov     qword, ptr, [rdi, +, 16], r10
 mov     qword, ptr, [rdi, +, 24], r11
```

</details>