@tarcieri 2022-08-08
Thanks! We've had some complains in the past about `#[inline(always)]` and its impact on compile times, but this seems like a justifiable use.